+++ src/commands/claim/claim.ts
@@ -1,8 +1,9 @@
 import { select } from '@inquirer/prompts';
-import { Command } from '@oclif/core';
+import { Command, Flags } from '@oclif/core';
 import { prompt } from '@oclif/core/lib/cli-ux/prompt.js';
 import { readFile, writeFile } from 'node:fs/promises';
 
+import { DATA_FOLDER } from '../../utils/constants.js';
 import { 
   claimCredentialOffer, 
   listFilesAsInquirerChoice, 
@@ -16,14 +17,31 @@ export default class Claim extends Command {
   static description = 'Claim Credential Offer from specified file';
 
   static examples = [
-    `$ oex claim`,
+    `$ oex claim --file <credential-offer.txt>`,
   ];
 
+  static flags = {
+    file: Flags.string({ 
+      char: 'f',
+      description: `credential offer filename in "${DATA_FOLDER}" folder`
+    })
+  };
+
   async run(): Promise<void> {
-    const offerFile = await select({
-      choices: listFilesAsInquirerChoice(".data"),
-      message: 'Select Credential Offer file'
-    });
+    const { flags } = await this.parse(Claim);
+
+    let offerFile;
+
+    if (flags.file) {
+      offerFile = `${DATA_FOLDER}/${flags.file}`
+    }
+
+    if (!offerFile) {
+      offerFile = await select({
+        choices: listFilesAsInquirerChoice(DATA_FOLDER),
+        message: 'Select Credential Offer file'
+      });
+    }
   
     const credentialOffer = await readFile(offerFile).then((data) => data.toString());
 
@@ -33,6 +51,6 @@ export default class Claim extends Command {
 
     const vcFilename = await prompt('Save Credential as', { default: prependTS('credential.jwt') });
 
-    await writeFile(`./.data/${vcFilename}`, verifiableCredential);
+    await writeFile(`${DATA_FOLDER}/${vcFilename}`, verifiableCredential);
   }
 }