+++ src/commands/present/present.ts
@@ -1,10 +1,11 @@
 import { select } from '@inquirer/prompts';
-import { Command, ux } from '@oclif/core';
+import { Command, Flags, ux } from '@oclif/core';
 import { decodeJwt } from 'jose';
 import { readFile, writeFile } from 'node:fs/promises';
 
 import type { PresentationRequest } from '../../types/index.js';
 
+import { DATA_FOLDER } from '../../utils/constants.js';
 import { 
   generatePresentationRequestSubmission, 
   listFilesAsInquirerChoice, 
@@ -21,17 +22,34 @@ export default class Present extends Command {
     `$ oex present`,
   ];
 
+  static flags = {
+    file: Flags.string({ 
+      char: 'f',
+      description: `presentation request filename in "${DATA_FOLDER}" folder`
+    })
+  };
+
   async run(): Promise<void> {
-    const requestFile = await select({
-      choices: listFilesAsInquirerChoice(".data"),
-      message: 'Select Presentation Request file'
-    });
+    const { flags } = await this.parse(Present);
+
+    let requestFile;
+
+    if (flags.file) {
+      requestFile = `${DATA_FOLDER}/${flags.file}`
+    }
+
+    if (!requestFile) {
+      requestFile = await select({
+        choices: listFilesAsInquirerChoice(DATA_FOLDER),
+        message: 'Select Presentation Request file'
+      });
+    }
 
     const presentrationRequest = await readFile(requestFile, { encoding: 'utf8' }).then((data) => data.toString());
     const presentationRequestURI = parsePresentationRequestURI(presentrationRequest);
 
     const credentialFile = await select({
-      choices: listFilesAsInquirerChoice('.data', /.jwt$/),
+      choices: listFilesAsInquirerChoice(DATA_FOLDER, /.jwt$/),
       message: 'Select Credential file'
     });
   
@@ -52,7 +70,7 @@ export default class Present extends Command {
 
     const submissionFile = prependTS('submission.json');
     this.log('Saving Presentation Request submission to', submissionFile);
-    await writeFile(`./.data/${submissionFile}`, JSON.stringify(requestSubmission));
+    await writeFile(`${DATA_FOLDER}/${submissionFile}`, JSON.stringify(requestSubmission));
     
     const response = await fetch(requestPayload.redirect_uri, {
       body: JSON.stringify(requestSubmission),
@@ -64,6 +82,6 @@ export default class Present extends Command {
     
     const submissionResultFile = prependTS('submission-result.json');
     this.log('Saving Presentation Request submission Result to', submissionResultFile);
-    writeFile(`./.data/${submissionResultFile}`, JSON.stringify(response));
+    writeFile(`${DATA_FOLDER}/${submissionResultFile}`, JSON.stringify(response));
   }
 }