+++ src/types/create-credential-offer.types.ts
@@ -1,8 +1,12 @@
 export type CredentialMetadata = {
+  credential_definition: {
+    types?: string[];
+    vct?: string;
+  };
   display: Array<{ name: string }>;
   format: string;
   id: string;
-  types: string[];
+  types?: string[];
 }
 
 export type Credential = {