+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.2.0] (2023.12.11)
 Add support to get `token_endpoint` and `grant_types_supported` from `.well-known/openid-configuration`
 Add [-f --file] flags to `present` and `claim` commands
+Add support for `credential_definition` in `credentials_supported`
 
 ## [1.1.0] (2023.12.07)
 Parse `credentials_supported` as map